+++ routes/authentication.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 
-const {signup, signin, signout, requireSignIn} = require("../controllers/user");
+const {signup, signin, signout, requireSignIn} = require("../controllers/authentication");
 const {UserSignUpValidator} = require('../helpers/validateUser');
 
 router.post('/signup', UserSignUpValidator, signup);
